#3 Rename login_control to login_info_control
